requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
gunicorn==21.2.0
djangorestframework==3.14.0
django-pgcrypto-fields==2.6.0
rules==3.3
django-cors-headers==4.3.1
python-dateutil==2.8.2
webauthn==2.0.0
#remove following line after removing Djoser code, if ever
djangorestframework-simplejwt==5.3.0
